requirements.txt
# Core dependencies
streamlit==1.28.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
pymysql==1.1.0
psycopg2-binary==2.9.9
pandas==2.1.4
groq==0.4.1

# Data visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Report generation
reportlab==4.0.7
jinja2==3.1.2
weasyprint==60.2

# Performance and optimization
sqlparse==0.4.4

# Session management
streamlit-authenticator==0.2.3

# Additional utilities
numpy==1.25.2